Extract PPQ provider panels
